requirements.txt
torch
numpy
pillow
requests
opencv-python
torchvision
matplotlib
flask